requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
httpx
email-validator
alembic